+++ YoutubeLike/Views/VideoLauncher.swift
@@ -3,83 +3,33 @@ import AVFoundation
 
 class VideoPlayerView: UIView {
     
-    let activityIndicatorView: UIActivityIndicatorView = {
-        let aiv = UIActivityIndicatorView(activityIndicatorStyle: UIActivityIndicatorViewStyle.whiteLarge)
-        aiv.translatesAutoresizingMaskIntoConstraints = false
-        aiv.startAnimating()
-        return aiv
-    }()
-    
-    let pausePlayButton: UIButton = {
-        let button = UIButton(type: UIButtonType.system)
-        let image = UIImage(named: "pause")
-        button.setImage(image, for: .normal)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.tintColor = UIColor.white
-        button.isHidden = true
-        button.addTarget(self, action: #selector(handlePausePlay), for: .touchUpInside)
-        return button
-    }()
-    
-    let controlContainerView: UIView = {
-        let view = UIView()
+    let controlContainerView: VideoControl = {
+        let view = VideoControl()
         view.backgroundColor = UIColor(white: 0, alpha: 0.5)
         return view
     }()
     
-    let videoLengthLabel: UILabel = {
-        let label = UILabel()
-        label.text = "00:00"
-        label.textColor = UIColor.white
-        label.font = UIFont.boldSystemFont(ofSize: 13)
-        label.textAlignment = .right
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let currentTimeLabel: UILabel = {
-        let label = UILabel()
-        label.text = "00:00"
-        label.textColor = UIColor.white
-        label.font = UIFont.boldSystemFont(ofSize: 13)
-        label.textAlignment = .right
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    let videoSlider: UISlider = {
-        let slider = UISlider()
-        slider.translatesAutoresizingMaskIntoConstraints = false
-        slider.minimumTrackTintColor = UIColor.red
-        slider.maximumTrackTintColor = UIColor.white
-        slider.setThumbImage(UIImage(named: "thumb"), for: .normal)
-        slider.addTarget(self, action: #selector(handleSliderChange), for: .valueChanged)
-        return slider
-    }()
-    
     var player: AVPlayer?
-    var isPlaying = false
     
     override init(frame: CGRect) {
         super.init(frame: frame)
         setupPlayerView()
-        setupGradientLayer()
         setupViews()
     }
     
     override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
         if keyPath == "currentItem.loadedTimeRanges" {
-            activityIndicatorView.stopAnimating()
-            controlContainerView.backgroundColor = UIColor.clear
-            pausePlayButton.isHidden = false
-            isPlaying = true
+//            activityIndicatorView.stopAnimating()
+//            controlContainerView.backgroundColor = UIColor.clear
+//            pausePlayButton.isHidden = false
+//            isPlaying = true
             
             guard let duration = player?.currentItem?.duration as? CMTime, let seconds = CMTimeGetSeconds(duration) as? Double else {
                 return
             }
             let minutesText = String(format: "%02d", Int(seconds) / 60)
             let secondsText = Int(seconds) % 60
-            videoLengthLabel.text = "\(minutesText):\(secondsText)"
+//            videoLengthLabel.text = "\(minutesText):\(secondsText)"
         }
     }
     
@@ -92,96 +42,61 @@ class VideoPlayerView: UIView {
         controlContainerView.frame = frame
         addSubview(controlContainerView)
         
-        controlContainerView.addSubview(activityIndicatorView)
-        activityIndicatorView.centerXAnchor.constraint(equalTo: centerXAnchor).isActive = true
-        activityIndicatorView.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
-        
-        controlContainerView.addSubview(pausePlayButton)
-        pausePlayButton.centerXAnchor.constraint(equalTo: centerXAnchor).isActive = true
-        pausePlayButton.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
-        pausePlayButton.widthAnchor.constraint(equalToConstant: 50).isActive = true
-        pausePlayButton.heightAnchor.constraint(equalToConstant: 50).isActive = true
-        
-        controlContainerView.addSubview(videoLengthLabel)
-        videoLengthLabel.rightAnchor.constraint(equalTo: rightAnchor, constant: -8).isActive = true
-        videoLengthLabel.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -2 ).isActive = true
-        videoLengthLabel.widthAnchor.constraint(equalToConstant: 60).isActive = true
-        videoLengthLabel.heightAnchor.constraint(equalToConstant: 24).isActive = true
-        
-        controlContainerView.addSubview(currentTimeLabel)
-        currentTimeLabel.leftAnchor.constraint(equalTo: leftAnchor, constant: -8).isActive = true
-        currentTimeLabel.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -2).isActive = true
-        currentTimeLabel.widthAnchor.constraint(equalToConstant: 60).isActive = true
-        currentTimeLabel.heightAnchor.constraint(equalToConstant: 24).isActive = true
-        
-        controlContainerView.addSubview(videoSlider)
-        videoSlider.rightAnchor.constraint(equalTo: videoLengthLabel.leftAnchor).isActive = true
-        videoSlider.leftAnchor.constraint(equalTo: currentTimeLabel.rightAnchor, constant: 8).isActive = true
-        videoSlider.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
-        videoSlider.heightAnchor.constraint(equalToConstant: 30).isActive = true
-        
-        
-        
         backgroundColor = UIColor.black
     }
     
-    func handlePausePlay() {
-        if isPlaying {
-            player?.pause()
-            pausePlayButton.setImage(UIImage(named: "play"), for: .normal)
-        } else {
-            player?.play()
-            pausePlayButton.setImage(UIImage(named: "pause"), for: .normal)
-        }
-        
-        isPlaying = !isPlaying
-    }
+//    func handlePausePlay() {
+//        if isPlaying {
+//            player?.pause()
+//            pausePlayButton.setImage(UIImage(named: "play"), for: .normal)
+//        } else {
+//            player?.play()
+//            pausePlayButton.setImage(UIImage(named: "pause"), for: .normal)
+//        }
+//        
+//        isPlaying = !isPlaying
+//    }
+//    
+//    func handleExpand() {
+//        
+//    }
     
-    func handleSliderChange() {
-        guard let duration: CMTime = player?.currentItem?.duration, let totalSeconds: Double = CMTimeGetSeconds(duration) else {
-            return
-        }
-        let value = Double(videoSlider.value) * totalSeconds
-        let seekTime = CMTime(value: Int64(value), timescale: 1)
-        player?.seek(to: seekTime, completionHandler: { (completed) in
-            
-        })
-    }
+//    func handleSliderChange() {
+//        guard let duration: CMTime = player?.currentItem?.duration, let totalSeconds: Double = CMTimeGetSeconds(duration) else {
+//            return
+//        }
+//        let value = Double(videoSlider.value) * totalSeconds
+//        let seekTime = CMTime(value: Int64(value), timescale: 1)
+//        player?.seek(to: seekTime, completionHandler: { (completed) in
+//            
+//        })
+//    }
     
     private func setupPlayerView() {
-        let urlString = "https://firebasestorage.googleapis.com/v0/b/gameofchats-762ca.appspot.com/o/message_movies%2F12323439-9729-4941-BA07-2BAE970967C7.mov?alt=media&token=3e37a093-3bc8-410f-84d3-38332af9c726"
-        guard let url = URL(string: urlString) else {
-            return
-        }
-        
-        player = AVPlayer(url: url)
-        let playerLayer = AVPlayerLayer(player: player)
-        self.layer.addSublayer(playerLayer)
-        playerLayer.frame = frame
-        
-        player!.play()
-        player!.addObserver(self, forKeyPath: "currentItem.loadedTimeRanges", options: .new, context: nil)
-        let interval = CMTime(value: 1, timescale: 2)
-        player!.addPeriodicTimeObserver(forInterval: interval, queue: DispatchQueue.main) { (progressTime) in
-            
-            let seconds = CMTimeGetSeconds(progressTime)
-            let secondsText = String(format: "%02d", Int(seconds) % 60)
-            let minutesText = String(format: "%02d", Int(seconds) / 60)
-            self.currentTimeLabel.text = "\(minutesText):\(secondsText)"
-
-            if let duration: CMTime = self.player?.currentItem?.duration, let totalSeconds: Double = CMTimeGetSeconds(duration) {
-                self.videoSlider.value = Float(seconds / totalSeconds)
-            }
-        }
-    }
-    
-    private func setupGradientLayer() {
-        let gradientLayer = CAGradientLayer()
-        gradientLayer.frame = bounds
-        gradientLayer.colors = [UIColor.clear.cgColor, UIColor.black.cgColor]
-        gradientLayer.locations = [0.7, 1.2]
-        
-        controlContainerView.layer.addSublayer(gradientLayer)
+//        let urlString = "https://firebasestorage.googleapis.com/v0/b/gameofchats-762ca.appspot.com/o/message_movies%2F12323439-9729-4941-BA07-2BAE970967C7.mov?alt=media&token=3e37a093-3bc8-410f-84d3-38332af9c726"
+//        guard let url = URL(string: urlString) else {
+//            return
+//        }
+//        
+//        player = AVPlayer(url: url)
+//        let playerLayer = AVPlayerLayer(player: player)
+//        self.layer.addSublayer(playerLayer)
+//        playerLayer.frame = frame
+//        
+//        player!.play()
+//        player!.addObserver(self, forKeyPath: "currentItem.loadedTimeRanges", options: .new, context: nil)
+//        let interval = CMTime(value: 1, timescale: 2)
+//        player!.addPeriodicTimeObserver(forInterval: interval, queue: DispatchQueue.main) { (progressTime) in
+//            
+//            let seconds = CMTimeGetSeconds(progressTime)
+//            let secondsText = String(format: "%02d", Int(seconds) % 60)
+//            let minutesText = String(format: "%02d", Int(seconds) / 60)
+//            self.currentTimeLabel.text = "\(minutesText):\(secondsText)"
+//
+//            if let duration: CMTime = self.player?.currentItem?.duration, let totalSeconds: Double = CMTimeGetSeconds(duration) {
+//                self.videoSlider.value = Float(seconds / totalSeconds)
+//            }
+//        }
     }
 }
